python-bridge/requirements.txt
asyncua>=1.0.0
websockets>=12.0
pyyaml>=6.0
